requirements.txt
google-adk==1.27.1
httpx==0.28.1
google-cloud-aiplatform==1.141.0

